student_package/requirements.txt
# Minimal requirements for students
opencv-python>=4.5.0
matplotlib>=3.5.0
numpy>=1.21.0
requests>=2.28.0
pillow>=9.0.0
